# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kido Learn Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  student: "student",
  teacher: "teacher",
} as const;

export interface RegisterRequest {
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
  firstName: string;
  lastName: string;
  role?: RegisterRequestRole;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: string;
  username: string;
  email: string;
  firstName: string;
  lastName: string;
  role: string;
  hasFace: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
  message: string;
}

export interface FaceRegisterRequest {
  /** 128-dimensional face descriptor from face-api.js */
  descriptor: number[];
}

export interface FaceVerifyRequest {
  /** 128-dimensional face descriptor to verify */
  descriptor: number[];
}

export interface FaceVerifyResponse {
  match: boolean;
  distance: number;
  message: string;
}

export type CourseDifficulty =
  (typeof CourseDifficulty)[keyof typeof CourseDifficulty];

export const CourseDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface Course {
  id: string;
  title: string;
  description: string;
  subject: string;
  gradeLevel: string;
  thumbnail?: string;
  totalLessons: number;
  durationMinutes: number;
  difficulty: CourseDifficulty;
  createdAt: string;
}

export interface Lesson {
  id: string;
  title: string;
  content: string;
  order: number;
  durationMinutes: number;
  videoUrl?: string;
}

export type CourseDetail = Course & {
  lessons: Lesson[];
};

export interface Progress {
  courseId: string;
  courseTitle: string;
  completedLessons: number;
  totalLessons: number;
  percentComplete: number;
  lastAccessedAt?: string;
}

export interface UpdateProgressRequest {
  lessonId: string;
  completed: boolean;
}

export type QuizRequestDifficulty =
  (typeof QuizRequestDifficulty)[keyof typeof QuizRequestDifficulty];

export const QuizRequestDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface QuizRequest {
  topic: string;
  numQuestions?: number;
  difficulty?: QuizRequestDifficulty;
}

export interface QuizQuestion {
  question: string;
  options: string[];
  correctIndex: number;
  explanation: string;
}

export interface QuizResponse {
  topic: string;
  questions: QuizQuestion[];
}

export interface RecommendationResponse {
  recommendations: Course[];
  reason: string;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}