# lib/api-zod/src/generated/types/course.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kido Learn Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CourseDifficulty } from "./courseDifficulty";

export interface Course {
  id: string;
  title: string;
  description: string;
  subject: string;
  gradeLevel: string;
  thumbnail?: string;
  totalLessons: number;
  durationMinutes: number;
  difficulty: CourseDifficulty;
  createdAt: string;
}