# lib/api-zod/src/generated/types/faceVerifyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kido Learn Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface FaceVerifyResponse {
  match: boolean;
  distance: number;
  message: string;
}