# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kido Learn Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterRequestRole } from "./registerRequestRole";

export interface RegisterRequest {
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
  firstName: string;
  lastName: string;
  role?: RegisterRequestRole;
}